# [multibody] Templatize body node