Templatizes BodyNodeImpl on concrete Mobilizer type to permit inlines rather than virtuals. This is part of the MbP performance epic #18442.

This PR doesn't change how we compute anything or attempt to reduce flop count. It just rearranges what's there to allow the compiler to generate better code. There are many further changes to come but I want to merge this now since the shuffling around would make it hard to review with substantive changes. This is a step towards a templatized kernel using only fixed-size objects, and reduces the amount of context-digging being done (but doesn't yet eliminate it all).

As a proof of concept, in addition to the newly-templatized BodyNodeImpl, I moved the position and velocity kinematics methods to take advantage of that. Everything else is unchanged. I measured 4% and 21% speedup for those with gcc, 11% and 32% with clang. Combined with the similar hacks to Frames in last week's #21853, position and velocity speedup is 28%/41% and 28%/44% resp. (See table below)
Note that these are the simplest computations with very few "up to 6" vectors and virtual function calls -- the dynamics computations have a lot more potential for speedup.

### What's here
- BodyNodeImpl (which was unused before) is now templated by concrete Mobilizer type, and partially filled in.
- Every mobilizer gains two new inline, non-virtual methods `calc_X_FM()` and `calc_V_FM()`. The existing virtual methods are reimplemented as calls to the non-virtuals.
- The per-node position & velocity kinematics functions are moved from the generic BodyNode to the templatized BodyNodeImpl where they can invoke the new inline methods directly.
- The corresponding tree-walking algorithms in MultibodyTree are fixed to extract numerical info from the context outside the loop, and to allow the individual nodes to grab what they need (which they can do very efficiently).

### What's not here
- no algorithmic changes
- no modifications to most of the existing computations (they are still stuck in BodyNode)
- not every Context extraction at low levels has been removed yet, so the existing "kernel" is not yet purely numerical
- other than modest speedups, there are no user-visible changes here (all code is `internal::`)

### Notes for reviewers
- I suggest starting with body_node_impl.h to see how it is templatized now, and the three virtual kinematics functions it implements. This is not CRTP -- it is very straightforward.
- Then body_node_impl.cc has the implementations which are able to call directly into concrete mobilizer's inlines. These functions were moved over directly from BodyNode and then lightly modified to replace virtuals with inline calls.
- Every mobilizer was changed in the same boilerplate manner to add the new inlines.

### Benchmarking
CassieBench measures only multibody kernel time, in microseconds. (I did nothing to speed up Mass Matrix calculation here but clang sped it up anyway so I included those times.)

|    Test (gcc)        |   Master    |  Master* | This PR | Speedup |  Speedup* |
| ------------------ | ----------- | ---------- | ------- | ------------ | ----------|
| Position only       |     2.68      |    3.58       |   2.58    |     4%       |     28%    |
| Position&Velocity|    5.55     |    7.52     |   4.41    |     21%        |    41%    |
| Mass Matrix         |   12.5     |    14.5     |   12.2    |     2%        |    16%    |

|    Test (clang)      |   Master    |  Master* | This PR | Speedup |  Speedup* |
| ------------------ | ----------- | ---------- | ------- | ------------ | ----------|
| Position only       |     2.52      |    3.12       |   2.25    |    11%       |     28%    |
| Position&Velocity|    5.19     |    6.3     |   3.55    |     32%        |    44%    |
| Mass Matrix        |    9.55     |    10.7     |  8.88    |     7%        |   17%    |

   `*` times from pre-#21853 which applied similar changes to Frames.

- Puget Xeon E5-2699 v4 @2.2GHz
- gcc 11.4.0
- clang 15.0.7

<!-- Reviewable:start -->
- - -
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/RobotLocomotion/drake/21862)
<!-- Reviewable:end -->