@sherm1 2024-09-10
Arggh! Ubsan steered me straight on this one -- it was an Eigen alignment issue. All even-sized fixed Eigen vectors require 16-byte alignment. So if you try to take a plain double array like our velocity state v, and cleverly overlay fixed Eigen vectors on it to do per-mobilizer math (as I was attempting), a `Vector1` works fine always but a `Vector6` seg faults if:
- the compiler generates SIMD instructions that have alignment restrictions, and
- the mobilizer's v segment happen to start at an odd index into the overall v.

Replaced for now with ugly but serviceable `std::array` objects instead :(