@sherm1 2024-09-10
Thanks to @jwnimmer-tri for helping me reproduce this locally.