@sherm1 2024-09-10
@drake-jenkins-bot mac-arm-ventura-clang-bazel-experimental-release please