@sherm1 2024-09-10
> Consider something like `Eigen::Matrix<nq, 1, Eigen::DontAlign>`

!!! That's just what I need. I went looking for that in Eigen's "alignment issues" documentation and didn't find it. Thanks! Will try.




<!-- Sent from Reviewable.io -->