@sherm1 2024-09-10
<details><summary><i>Previously, jwnimmer-tri (Jeremy Nimmer) wrote…</i></summary><blockquote>

Well, it's not failing anymore that CI can tell, so I'll mark this discussion resolved.  I am still extremely skeptical about using `reinterpret_cast` in anger, though.  It will probably bite us later with new platforms / compilers.  There are better and more conventional spellings of these API signatures that would avoid embracing UB.  I'll let you try to convince the platform reviewer, though, I'm happy to let someone else have first crack at that fight.
</blockquote></details>

UBsan gives the current scheme a clean bill of health. Maybe it's legit?




<!-- Sent from Reviewable.io -->